Add files via upload
The life of humans living without liver tumors is one of the fundamental cares of human livelihood. Therefore, for better care, detection of liver disease at a primitive phase is necessary. For medical experts, predicting the illness in the early stages due to subtle signs is a very difficult task. Many, when it is too late, the signs become evident. The current work aims to augment the perceived nature of liver disease using machine learning methods to solve this epidemic.